Trimming more than 500 stories from feeds. One day these can remain, but for now, it's too much to handle.

+++ apps/rss_feeds/models.py
@@ -417,7 +417,8 @@ class Feed(models.Model):
             story_feed_id=self.pk,
         ).order_by('-story_date')[500:]
         print 'Found %s extra stories in %s. Trimming...' % (stories.count(), self)
-        stories.delete()
+        if stories.count() > 500:
+            stories.delete()
         
     def get_stories(self, offset=0, limit=25, force=False):
         stories = cache.get('feed_stories:%s-%s-%s' % (self.id, offset, limit), [])
